Edit configuration files. Issue #31
