+++ src/main/java/dpfmanager/shell/modules/classes/Rules.java
@@ -39,6 +39,7 @@ public class Rules {
    * @param scene the scene
    */
   public void Read(Scene scene) {
+    rules.clear();
     Boolean wrong_format = false;
     AnchorPane ap2 = (AnchorPane)scene.lookup("#pane1");
     for (Node node : ap2.getChildren()){
