+++ src/main/resources/fxml/design.fxml
@@ -89,17 +89,22 @@
                                                         <Cursor fx:constant="TEXT" />
                                                     </cursor>
                                                 </TextField>
-                                                <Button layoutX="255.0" layoutY="649.0" mnemonicParsing="false" onAction="#openConfig" prefHeight="44.0" prefWidth="160.0" stylesheets="@/styles/style.css" text="IMPORT" textFill="#31C2F3" visible="true">
+                                                <Button layoutX="255.0" layoutY="634.0" mnemonicParsing="false" onAction="#openConfig" prefHeight="30.0" prefWidth="243.0" stylesheets="@/styles/style.css" text="IMPORT" textFill="#31C2F3" visible="true">
                                                     <cursor>
                                                         <Cursor fx:constant="HAND" />
                                                     </cursor>
                                                 </Button>
-                                                <Button layoutX="421.0" layoutY="649.0" mnemonicParsing="false" onAction="#newConfig" prefHeight="45.0" prefWidth="155.0" stylesheets="@/styles/style.css" text="NEW" textFill="#31C2F3" visible="true">
+                                                <Button layoutX="505.0" layoutY="634.0" mnemonicParsing="false" onAction="#newConfig" prefHeight="30.0" prefWidth="235.0" stylesheets="@/styles/style.css" text="NEW" textFill="#31C2F3" visible="true">
                                                     <cursor>
                                                         <Cursor fx:constant="HAND" />
                                                     </cursor>
                                                 </Button>
-                                                <Button layoutX="582.0" layoutY="649.0" mnemonicParsing="false" onAction="#deleteConfig" prefHeight="45.0" prefWidth="155.0" stylesheets="@/styles/style.css" text="DELETE" textFill="#31C2F3" visible="true">
+                                                <Button layoutX="255.0" layoutY="671.0" mnemonicParsing="false" onAction="#editConfig" prefHeight="30.0" prefWidth="243.0" stylesheets="@/styles/style.css" text="EDIT" textFill="#31C2F3" visible="true">
+                                                    <cursor>
+                                                        <Cursor fx:constant="HAND" />
+                                                    </cursor>
+                                                </Button>
+                                                <Button layoutX="505.0" layoutY="671.0" mnemonicParsing="false" onAction="#deleteConfig" prefHeight="30.0" prefWidth="235.0" stylesheets="@/styles/style.css" text="DELETE" textFill="#31C2F3" visible="true">
                                                     <cursor>
                                                         <Cursor fx:constant="HAND" />
                                                     </cursor>
